requirements.txt
# 核心深度学习框架
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 科学计算和数据处理
numpy>=1.21.0
pandas>=1.5.0
scipy>=1.9.0

# 计算机视觉
opencv-python>=4.5.0
Pillow>=9.0.0

# 机器学习
scikit-learn>=1.1.0

# 数据可视化
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Web框架
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6

# 数据处理和序列化
pydantic>=2.0.0
python-json-logger>=2.0.0

# 工具库
tqdm>=4.64.0
requests>=2.28.0
python-dotenv>=0.19.0
pathlib2>=2.3.0

# 开发和测试
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# 性能监控
psutil>=5.9.0
GPUtil>=1.4.0

# 文档生成
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# 中文支持
matplotlib-fontproperties>=0.0.1